refactor: split schema regen and version bump into separate scripts
- Rename Fleet-Control-Schema.yml -> Agent-Config-Schema.yml (workflow regenerates schema only)
- Rename fleet-control-schema.gradle -> agent-config-schema.gradle
- Add BumpSchemaVersion.groovy: release-time bump (--since=<ref> --ci) opens a PR
- Add SchemaDiff.groovy: shared diff/bump helpers (classifyChanges, recommendBump, applyBump, bumpVersion)
- Add Agent-Config-Schema-Bump.yml: workflow_run trigger on Test Suite - Release success
- Slim GenerateSchema.groovy: regen-only, no version writes; loads SchemaDiff via GroovyShell
- Fix extractComments: replace Groovy List.pop() (removes FRONT, not back) with explicit index removal — Java-version-safe and stack-correct
- Split tests: GenerateSchemaTest covers regen, SchemaDiffTest covers diff/bump
