+++ .github/workflows/Agent-Config-Schema.yml
@@ -1,19 +1,21 @@
-name: Fleet Control Config Schema
+name: Agent Config Schema
+
+# Permissions granted to all jobs in the workflow.
+# Permissions that aren't explicitly defined here will default to a value of none.
+permissions: {}
 
 # Regenerates .fleetControl/schemas/config.json from reference-newrelic.yml
 # whenever a push touches relevant inputs, then commits the regenerated schema
-# (and any version bump in .fleetControl/configurationDefinitions.yml) back to
-# the pushed branch.
+# back to the pushed branch.
+#
+# Version bumps to .fleetControl/configurationDefinitions.yml are NOT done
+# here — they happen at release time via Agent-Config-Schema-Bump.yml.
 #
 # Also fails the current push if newrelic.yml changed but reference-newrelic.yml
 # didn't. The check is per-push, so a follow-up push that updates
 # reference-newrelic.yml will pass and re-green the PR.
 #
-# Skips main because branch protection blocks the bot from pushing there. The
-# drift check runs on the feature branch before merge, so drift can't reach
-# main without going through a feature-branch run first.
-
-permissions: {}
+# Skips main because branch protection blocks the bot from pushing there.
 
 on:
   push:
@@ -24,7 +26,7 @@ on:
       - '.fleetControl/schemaGeneration/**'
 
 concurrency:
-  group: fleet-control-schema-${{ github.ref }}
+  group: agent-config-schema-${{ github.ref }}
   cancel-in-progress: true
 
 jobs:
@@ -35,7 +37,8 @@ jobs:
       contents: write
 
     steps:
-      - uses: actions/checkout@0ad4b8fadaa221de15dcec353f45205ec38ea70b  # pin@v4
+#      - uses: GitHubSecurityLab/actions-permissions/monitor@v1 # detects minimum permissions required to run the workflow, uncomment to use
+      - uses: actions/checkout@0ad4b8fadaa221de15dcec353f45205ec38ea70b # pin@v4
         with:
           fetch-depth: 0
           token: ${{ secrets.GITHUB_TOKEN }}
@@ -57,9 +60,9 @@ jobs:
           fi
 
       - name: Set up Java
-        uses: actions/setup-java@v4
+        uses: actions/setup-java@b36c23c0d998641eff861008f374ee103c25ac73 # pin@v4
         with:
-          distribution: temurin
+          distribution: 'temurin'
           java-version: '17'
 
       - name: Install Groovy
@@ -71,7 +74,7 @@ jobs:
         id: generate
         run: |
           set +e
-          groovy .fleetControl/schemaGeneration/GenerateSchema.groovy --ci
+          groovy .fleetControl/schemaGeneration/GenerateSchema.groovy
           code=$?
           set -e
           case "$code" in
@@ -83,12 +86,12 @@ jobs:
       - name: Commit and push regenerated schema
         if: steps.generate.outputs.changed == 'true'
         run: |
-          if [ -z "$(git status --porcelain .fleetControl)" ]; then
-            echo "Generator reported changes but working tree is clean — nothing to commit."
+          if [ -z "$(git status --porcelain .fleetControl/schemas/config.json)" ]; then
+            echo "Generator reported changes but config.json is clean — nothing to commit."
             exit 0
           fi
           git config user.name 'github-actions[bot]'
           git config user.email '41898282+github-actions[bot]@users.noreply.github.com'
-          git add .fleetControl/schemas/config.json .fleetControl/configurationDefinitions.yml
-          git commit -m "chore: regenerate Fleet Control config schema"
+          git add .fleetControl/schemas/config.json
+          git commit -m "chore: regenerate agent config schema"
           git push
