+++ build.gradle
@@ -77,4 +77,4 @@ ext {
     asmVersion="9.9.1"
 }
 
-apply from: "$rootProject.projectDir/gradle/script/fleet-control-schema.gradle"
+apply from: "$rootProject.projectDir/gradle/script/agent-config-schema.gradle"
